requirements.txt
# WatchlistBot Dependencies
# Install with: pip install -r requirements.txt

# Core
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
pyyaml>=6.0

# Scheduling
schedule>=1.2.0
pytz>=2024.1

# AI scoring
anthropic>=0.25.0

# Optional: Reddit API (PRAW)
# praw>=7.7.0

# Optional: Technical analysis helpers
# ta-lib         # requires C library, see https://ta-lib.org
